+++ src/e2ee/worker/FrameCryptor.ts
@@ -8,7 +8,7 @@ import { ENCRYPTION_ALGORITHM, IV_LENGTH, UNENCRYPTED_BYTES } from '../constants
 import { CryptorError, CryptorErrorReason } from '../errors';
 import { CryptorCallbacks, CryptorEvent } from '../events';
 import type { DecodeRatchetOptions, KeyProviderOptions, KeySet } from '../types';
-import { deriveKeys, isVideoFrame } from '../utils';
+import { deriveKeys, isVideoFrame, needsRbspUnescaping, parseRbsp, writeRbsp } from '../utils';
 import type { ParticipantKeyHandler } from './ParticipantKeyHandler';
 import { SifGuard } from './SifGuard';
 
@@ -211,13 +211,9 @@ export class FrameCryptor extends BaseFrameCryptor {
         encodedFrame.getMetadata().synchronizationSource ?? -1,
         encodedFrame.timestamp,
       );
-
+      let frameInfo = this.getUnencryptedBytes(encodedFrame);
       // Thіs is not encrypted and contains the VP8 payload descriptor or the Opus TOC byte.
-      const frameHeader = new Uint8Array(
-        encodedFrame.data,
-        0,
-        this.getUnencryptedBytes(encodedFrame),
-      );
+      const frameHeader = new Uint8Array(encodedFrame.data, 0, frameInfo.unencryptedBytes);
 
       // Frame trailer contains the R|IV_LENGTH and key index
       const frameTrailer = new Uint8Array(2);
@@ -240,20 +236,25 @@ export class FrameCryptor extends BaseFrameCryptor {
             additionalData: new Uint8Array(encodedFrame.data, 0, frameHeader.byteLength),
           },
           encryptionKey,
-          new Uint8Array(encodedFrame.data, this.getUnencryptedBytes(encodedFrame)),
+          new Uint8Array(encodedFrame.data, frameInfo.unencryptedBytes),
         );
 
-        const newData = new ArrayBuffer(
-          frameHeader.byteLength + cipherText.byteLength + iv.byteLength + frameTrailer.byteLength,
+        let newDataWithoutHeader = new Uint8Array(
+          cipherText.byteLength + iv.byteLength + frameTrailer.byteLength,
         );
-        const newUint8 = new Uint8Array(newData);
+        newDataWithoutHeader.set(new Uint8Array(cipherText)); // add ciphertext.
+        newDataWithoutHeader.set(new Uint8Array(iv), cipherText.byteLength); // append IV.
+        newDataWithoutHeader.set(frameTrailer, cipherText.byteLength + iv.byteLength); // append frame trailer.
 
-        newUint8.set(frameHeader); // copy first bytes.
-        newUint8.set(new Uint8Array(cipherText), frameHeader.byteLength); // add ciphertext.
-        newUint8.set(new Uint8Array(iv), frameHeader.byteLength + cipherText.byteLength); // append IV.
-        newUint8.set(frameTrailer, frameHeader.byteLength + cipherText.byteLength + iv.byteLength); // append frame trailer.
+        if (frameInfo.isH264) {
+          newDataWithoutHeader = writeRbsp(newDataWithoutHeader);
+        }
 
-        encodedFrame.data = newData;
+        var newData = new Uint8Array(frameHeader.byteLength + newDataWithoutHeader.byteLength);
+        newData.set(frameHeader);
+        newData.set(newDataWithoutHeader, frameHeader.byteLength);
+
+        encodedFrame.data = newData.buffer;
 
         return controller.enqueue(encodedFrame);
       } catch (e: any) {
@@ -291,6 +292,10 @@ export class FrameCryptor extends BaseFrameCryptor {
       this.sifGuard.recordSif();
 
       if (this.sifGuard.isSifAllowed()) {
+        encodedFrame.data = encodedFrame.data.slice(
+          0,
+          encodedFrame.data.byteLength - this.sifTrailer.byteLength,
+        );
         return controller.enqueue(encodedFrame);
       } else {
         workerLogger.warn('SIF limit reached, dropping frame');
@@ -321,11 +326,11 @@ export class FrameCryptor extends BaseFrameCryptor {
       }
     } else if (!this.keys.getKeySet(keyIndex) && this.keys.hasValidKey) {
       // emit an error in case the key index is out of bounds but the key handler thinks we still have a valid key
-      workerLogger.warn('skipping decryption due to missing key at index');
+      workerLogger.warn(`skipping decryption due to missing key at index ${keyIndex}`);
       this.emit(
         CryptorEvent.Error,
         new CryptorError(
-          `missing key at index for participant ${this.participantIdentity}`,
+          `missing key at index ${keyIndex} for participant ${this.participantIdentity}`,
           CryptorErrorReason.MissingKey,
         ),
       );
@@ -346,7 +351,7 @@ export class FrameCryptor extends BaseFrameCryptor {
     if (!ratchetOpts.encryptionKey && !keySet) {
       throw new TypeError(`no encryption key found for decryption of ${this.participantIdentity}`);
     }
-
+    let frameInfo = this.getUnencryptedBytes(encodedFrame);
     // Construct frame trailer. Similar to the frame header described in
     // https://tools.ietf.org/html/draft-omara-sframe-00#section-4.2
     // but we put it at the end.
@@ -356,11 +361,20 @@ export class FrameCryptor extends BaseFrameCryptor {
     // ---------+-------------------------+-+---------+----
 
     try {
-      const frameHeader = new Uint8Array(
+      const frameHeader = new Uint8Array(encodedFrame.data, 0, frameInfo.unencryptedBytes);
+      var encryptedData = new Uint8Array(
         encodedFrame.data,
-        0,
-        this.getUnencryptedBytes(encodedFrame),
+        frameHeader.length,
+        encodedFrame.data.byteLength - frameHeader.length,
       );
+      if (frameInfo.isH264 && needsRbspUnescaping(encryptedData)) {
+        encryptedData = parseRbsp(encryptedData);
+        const newUint8 = new Uint8Array(frameHeader.byteLength + encryptedData.byteLength);
+        newUint8.set(frameHeader);
+        newUint8.set(encryptedData, frameHeader.byteLength);
+        encodedFrame.data = newUint8.buffer;
+      }
+
       const frameTrailer = new Uint8Array(encodedFrame.data, encodedFrame.data.byteLength - 2, 2);
 
       const ivLength = frameTrailer[0];
@@ -404,7 +418,7 @@ export class FrameCryptor extends BaseFrameCryptor {
           );
 
           let ratchetedKeySet: KeySet | undefined;
-          if (keySet === this.keys.getKeySet(keyIndex)) {
+          if ((initialMaterial ?? keySet) === this.keys.getKeySet(keyIndex)) {
             // only ratchet if the currently set key is still the same as the one used to decrypt this frame
             // if not, it might be that a different frame has already ratcheted and we try with that one first
             const newMaterial = await this.keys.ratchetKey(keyIndex, false);
@@ -417,22 +431,21 @@ export class FrameCryptor extends BaseFrameCryptor {
             encryptionKey: ratchetedKeySet?.encryptionKey,
           });
           if (frame && ratchetedKeySet) {
-            this.keys.setKeySet(ratchetedKeySet, keyIndex, true);
-            // decryption was successful, set the new key index to reflect the ratcheted key set
-            this.keys.setCurrentKeyIndex(keyIndex);
+            // before updating the keys, make sure that the keySet used for this frame is still the same as the currently set key
+            // if it's not, a new key might have been set already, which we don't want to override
+            if ((initialMaterial ?? keySet) === this.keys.getKeySet(keyIndex)) {
+              this.keys.setKeySet(ratchetedKeySet, keyIndex, true);
+              // decryption was successful, set the new key index to reflect the ratcheted key set
+              this.keys.setCurrentKeyIndex(keyIndex);
+            }
           }
           return frame;
         } else {
           /**
-           * Since the key it is first send and only afterwards actually used for encrypting, there were
-           * situations when the decrypting failed due to the fact that the received frame was not encrypted
-           * yet and ratcheting, of course, did not solve the problem. So if we fail RATCHET_WINDOW_SIZE times,
-           * we come back to the initial key.
+           * Because we only set a new key once decryption has been successful,
+           * we can be sure that we don't need to reset the key to the initial material at this point
+           * as the key has not been updated on the keyHandler instance
            */
-          if (initialMaterial) {
-            workerLogger.debug('resetting to initial material');
-            this.keys.setKeyFromMaterial(initialMaterial.material, keyIndex);
-          }
 
           workerLogger.warn('maximum ratchet attempts exceeded');
           throw new CryptorError(
@@ -489,7 +502,11 @@ export class FrameCryptor extends BaseFrameCryptor {
     return iv;
   }
 
-  private getUnencryptedBytes(frame: RTCEncodedVideoFrame | RTCEncodedAudioFrame): number {
+  private getUnencryptedBytes(frame: RTCEncodedVideoFrame | RTCEncodedAudioFrame): {
+    unencryptedBytes: number;
+    isH264: boolean;
+  } {
+    var frameInfo = { unencryptedBytes: 0, isH264: false };
     if (isVideoFrame(frame)) {
       let detectedCodec = this.getVideoCodec(frame) ?? this.videoCodec;
 
@@ -498,7 +515,8 @@ export class FrameCryptor extends BaseFrameCryptor {
       }
 
       if (detectedCodec === 'vp8') {
-        return UNENCRYPTED_BYTES[frame.type];
+        frameInfo.unencryptedBytes = UNENCRYPTED_BYTES[frame.type];
+        return frameInfo;
       }
 
       const data = new Uint8Array(frame.data);
@@ -506,19 +524,20 @@ export class FrameCryptor extends BaseFrameCryptor {
         const naluIndices = findNALUIndices(data);
 
         // if the detected codec is undefined we test whether it _looks_ like a h264 frame as a best guess
-        const isH264 =
+        frameInfo.isH264 =
           detectedCodec === 'h264' ||
           naluIndices.some((naluIndex) =>
             [NALUType.SLICE_IDR, NALUType.SLICE_NON_IDR].includes(parseNALUType(data[naluIndex])),
           );
 
-        if (isH264) {
+        if (frameInfo.isH264) {
           for (const index of naluIndices) {
             let type = parseNALUType(data[index]);
             switch (type) {
               case NALUType.SLICE_IDR:
               case NALUType.SLICE_NON_IDR:
-                return index + 2;
+                frameInfo.unencryptedBytes = index + 2;
+                return frameInfo;
               default:
                 break;
             }
@@ -529,9 +548,11 @@ export class FrameCryptor extends BaseFrameCryptor {
         // no op, we just continue and fallback to vp8
       }
 
-      return UNENCRYPTED_BYTES[frame.type];
+      frameInfo.unencryptedBytes = UNENCRYPTED_BYTES[frame.type];
+      return frameInfo;
     } else {
-      return UNENCRYPTED_BYTES.audio;
+      frameInfo.unencryptedBytes = UNENCRYPTED_BYTES.audio;
+      return frameInfo;
     }
   }